@@ -11,7 +11,7 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = "~3.9"
+python = "^3.10"
 aiohttp = {version = "^3.8.3", extras = ["speedups"]}
 aiofiles = "^22.1.0"
 aiocsv = "^1.2.3"
